feat: add clipboard parser
